Fix package command in makefile.
As previously written, the package rule would fail if a dist folder did not already
exist.  Since the dist the folder is not tracked by git, anyone who
clones the repo who runs `make package` will fail.  This change causes
the dist folder to only be cleared out if the dist folder already
exists.

+++ Makefile
@@ -14,7 +14,7 @@ doc/examples.md: scripts/exampleDoc.py
 	PYTHONPATH=`pwd` python3 scripts/exampleDoc.py > $@
 
 package:
-	rm dist/*
+	test -d dist && rm dist/* || mkdir dist/
 	python3 setup.py sdist bdist_wheel
 
 install: package
